Backend/requirements.txt
flask
flask_cors
pyyaml